2751. Iota 20
FIX: PatchUnitsLimit: Crash with this option due to overlapping of patching addresses between different patches. To diagnose such intersections in the future, a MemWatchDog module has been added.
FIX: PatchResetEngineersOrder: With the setting turned OFF, the original algorithm was broken, which led to an overestimation of the engineer's work counter when joined to another worker. The error has been present since 1.16. I hope no one has taken advantage of this or suffered from it.